some moving around of filenames and test fns
