+++ phylo/src/tkf_model/reestimate.rs
@@ -211,7 +211,7 @@ impl<'a, T: TKFModel, AA: AncestralAlignment> EdgeSeqsReestimator<'a, T, AA> {
         }
     }
 
-    pub(super) fn prepare_for_dp(&mut self, v2_idx: &NodeIdx) {
+    pub fn prepare_for_dp(&mut self, v2_idx: &NodeIdx) {
         let num_blocks = self.cost.model_info.borrow().blocks.len();
         self.dp_table = vec![[f64::NEG_INFINITY; DP_ASSIGNMENT_AND_EVENTS_SIZE]; num_blocks];
         //self.backtracking_table = vec![[0; DP_ASSIGNMENT_AND_EVENTS_SIZE]; num_blocks];
@@ -625,7 +625,7 @@ pub fn mapping_from_node_seq(node_seq: &NodeSeq, block_lens: &[usize], seq_len: 
     mapping
 }
 
-pub(super) fn possibilities_for_choices(
+fn possibilities_for_choices(
     choices: &[usize],
     no_choices: &[bool; N_EDGES_IN_QUARTET],
 ) -> QuartetDelOrNotPossibilities {
@@ -690,6 +690,69 @@ mod private_tests {
     use crate::phylo_info::PhyloInfoBuilder;
     use crate::tkf_model::tests::setup_test_phylo;
     use crate::tkf_model::{EdgeSeqsReestimator, TKF92IndelCostBuilder};
+
+    use super::*;
+
+    #[test]
+    fn tkf_reestimate_possibilities_for_choices_no_choice() {
+        let no_choices = [true, false, false, true, false];
+        let choices = vec![];
+        let possibilities = possibilities_for_choices(&choices, &no_choices);
+        let expected = vec![[true, false, false, true, false]];
+        assert_eq!(possibilities, expected);
+    }
+
+    #[test]
+    fn tkf_reestimate_possibilities_for_choices_one() {
+        let no_choices = [true, false, false, true, false];
+        let choices = vec![0];
+        let mut possibilities = possibilities_for_choices(&choices, &no_choices);
+        let mut expected = vec![
+            [true, false, false, true, false],
+            [false, false, false, true, false],
+        ];
+        // sort
+        possibilities.sort();
+        expected.sort();
+        assert_eq!(possibilities, expected);
+    }
+
+    #[test]
+    fn tkf_reestimate_possibilities_for_choices() {
+        let no_choices = [true, false, false, true, false];
+        let choices = vec![1, 3];
+        let mut possibilities = possibilities_for_choices(&choices, &no_choices);
+        let mut expected = vec![
+            [true, false, false, false, false],
+            [true, false, false, true, false],
+            [true, true, false, false, false],
+            [true, true, false, true, false],
+        ];
+        // sort
+        possibilities.sort();
+        expected.sort();
+        assert_eq!(possibilities, expected);
+    }
+
+    #[test]
+    fn tkf_reestimate_possibilities_for_choices_all() {
+        let no_choices = [true, false, false, true, false];
+        let choices = vec![0, 1, 2, 3, 4];
+        let possibilities = possibilities_for_choices(&choices, &no_choices);
+        let expected = (0..32)
+            .map(|i| {
+                [
+                    (i & 0b00001) != 0,
+                    (i & 0b00010) != 0,
+                    (i & 0b00100) != 0,
+                    (i & 0b01000) != 0,
+                    (i & 0b10000) != 0,
+                ]
+            })
+            .collect::<Vec<[bool; 5]>>();
+
+        assert_eq!(possibilities, expected);
+    }
 
     #[test]
     fn tkf_remove_and_add_back_quartet() {
